requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
mysql-connector-python==8.0.32
protobuf==3.20.3
pytz==2022.7.1
requests==2.28.2
urllib3==1.26.14
